feat(testing): share the film-stage contract across workstation journeys (#16501)
